refactor(FileName): Extract logic to split name into its own function
